backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
watchdog>=4.0.0
pillow>=10.0.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
alembic>=1.13.0
